requirements.txt
langchain
langchain-google-genai
langchain_cohere
langchain_community
python-dotenv
streamlit==1.22.0
unstructured==0.9.2
tiktoken==0.4.0
faiss-cpu==1.7.4
libmagic
python-magic
python-magic-bin
streamlit-extras